pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2html-api"
version = "0.1.0"
description = "Convert PDF pages to HTML using OpenAI Vision API via HTTP"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "PDF2HTML API", email = "dev@example.com"}
]
keywords = ["pdf", "html", "openai", "vision", "api", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
pdf2html-api = "pdf2html_api.main:app"

[project.urls]
Homepage = "https://github.com/example/pdf2html-api"
Repository = "https://github.com/example/pdf2html-api"
Issues = "https://github.com/example/pdf2html-api/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src" 